Refactor jagex service
